Basic text-file I/O added
This is the last version updated as a Teaching Assistant which may be the
last version for a while unless I find a purpose to use it elsewhere or
there is demand by the community.

In this version, I've added basic text-file I/O functions for reading and
writing files.  The intention was to be used for setting up my TA scripts
which operate using CSVs for data sources.

Miscellaneous changes include small bugfixes in Iterator, basic command
prompt input, execution of programs, and the ability to force the
interpreter to quit early.