Added "config" parameter and re-implemented hash tables
Instead of using pairs of TStringList, I implemented a hash table in
Delphi (my version doesn't support TDictionary) and redesigned the
structure for the object and string hash tables to extend an abstract hash
table class.

The "@config" preprocessor statement is a write-only hash table (similar
to variables) where the purpose is solely to interact with the
interpreter.  One such param is "StringLengthLimit" which prevents users
from blowing through memory by concatenating a string to itself over and
over and exhausting resources.  SMTP mail settings has moved to "config"
as well as HTTP authentication.

I also implemented (but haven't successfully tested) proxy settings for
pointing the HTTP client in Euphora to a proxy server.  I believe the
issue may be related to something in ICS or possibly running a command
prior to the request -- I'll need to flip through the source to figure
this one out.